fix(lint): split sync/mod_tests.rs under the 500-line gate
